% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulForestPlot.R
\name{cumulforest}
\alias{cumulforest}
\title{cumulative Forest plot}
\usage{
cumulforest(yi, vi, measure, d, effectName = "Effect")
}
\arguments{
\item{yi}{A \code{string} of the variable which holds the vector of length k with the observed effect sizes or outcomes in the selected dataset (d)}

\item{vi}{A \code{string} of the variable which holds the vector of length k with the corresponding sampling variances in the selected dataset (d)}

\item{measure}{A character string indicating underlying summary measure.}

\item{d}{A \code{string} representing the dataset name that should be used for fitting.}
}
\value{
creates a cumulative forest plot
also creates a json file (imgHeight.json) that is used in a later api call to define the height of the plots
}
\description{
Using metafor rma.uni function to fit a meta-analytic multivariate/multilevel fixed- and random/mixed-effects model for the given dataset.
The model then is used as input for the metaviz viz_forest function to print a forrest plot
See Documentation of metafor and metaviz packages for details.
}
\author{
Robert Studtrucker
}
