% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestPlot.R
\name{forest}
\alias{forest}
\title{Forest plot}
\usage{
forest(yi, vi, measure, d, effect = "Effect")
}
\arguments{
\item{yi}{A \code{string} of the variable which holds the vector of length k with the observed effect sizes or outcomes in the selected dataset (d)}

\item{vi}{A \code{string} of the variable which holds the vector of length k with the corresponding sampling variances in the selected dataset (d)}

\item{measure}{A character string indicating underlying summary measure.}

\item{d}{A \code{string} representing the dataset name that should be used for fitting.}
}
\value{
creates a forest and rainforest plot
also creates a json file (imgHeight.json) that is used in a later api call to define the height of the plots
}
\description{
Using metafor package to create a forest plot and metaviz package to create a rain forest plot.
}
\author{
Robert Studtrucker
}
