% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnelPlot.R
\name{funnelPLot}
\alias{funnelPLot}
\title{Funnel plot}
\usage{
funnelPLot(yi, vi, measure, d, effectName = "Effect")
}
\arguments{
\item{yi}{A \code{string} of the variable which holds the vector of length k with the observed effect sizes or outcomes in the selected dataset (d)}

\item{vi}{A \code{string} of the variable which holds the vector of length k with the corresponding sampling variances in the selected dataset (d)}

\item{measure}{A character string indicating underlying summary measure.}

\item{d}{A \code{string} representing the dataset name that should be used for fitting.}
}
\value{
returns a Funnel plot for the given dataset
}
\description{
Using metafor package to create a funnel plot.
}
\author{
Robert Studtrucker
}
