% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{jsonSummary}
\alias{jsonSummary}
\title{JSON summary}
\usage{
jsonSummary(d)
}
\arguments{
\item{d}{A \code{string} representing the dataset name which json representation should be returned.}
}
\value{
returns a json representation of an included dataset
}
\description{
produces a json summary of includes dataset specified by parameter [d].
}
\examples{
jsonSummary("CAMA_Math")
}
\author{
Robert Studtrucker
}
