% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapower.R
\name{metapower}
\alias{metapower}
\title{meta power plot}
\usage{
metapower(yi, vi, measure, d, pval = 0.05)
}
\arguments{
\item{yi}{A \code{string} of the variable which holds the vector of length k with the observed effect sizes or outcomes in the selected dataset (d)}

\item{vi}{A \code{string} of the variable which holds the vector of length k with the corresponding sampling variances in the selected dataset (d)}

\item{measure}{A character string indicating underlying summary measure}

\item{d}{A \code{string} representing the dataset name.}

\item{pval}{A \code{Numeric} representing the significance level which should be used for the power simulation.}
}
\value{
A metapower plot
}
\description{
creates a metapower plot.
Not used at the moment
}
\author{
Robert Studtrucker
}
