% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMetaForestplot.R
\name{netMetaForestplot}
\alias{netMetaForestplot}
\title{netmeta forest plot}
\usage{
netMetaForestplot(model, reference)
}
\arguments{
\item{model}{A netmeta model object}

\item{reference}{A \code{string} representing the reference category that should be used. Here the reference is just needed for the ploting function.}
}
\value{
returns a forest plot for a given netmeta model.
}
\description{
Using the R package netmeta to to create and and return a forest plot for the given model and reference category
This indirect calling is needed since calling functions other than those from psychOpenCama package is blocked on the opencpu server for security reasons.
}
\author{
Robert Studtrucker
}
