% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMetaGetTRTS.R
\name{netMetagetTRTS}
\alias{netMetagetTRTS}
\title{Helper Function to get possible reference categories fot the netmeta model}
\usage{
netMetagetTRTS(d)
}
\arguments{
\item{d}{A \code{string} representing the dataset name for which the reference categories should be retrieved}
}
\value{
create a json file containing the reference categories in the opencpu session which can then be retrieved in another call
}
\description{
Helper Function to get possible reference categories to be used in the netmeta model
This function is used to populate the gui with reference options
}
\author{
Robert Studtrucker
}
