% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMetaMatrix.R
\name{netMetaMatrix}
\alias{netMetaMatrix}
\title{Matrix comparison}
\usage{
netMetaMatrix(model)
}
\arguments{
\item{model}{A netmeta model object}
}
\value{
returns a matrix of all pairwise comparison for the given netmeta model.
}
\description{
Using the R package netmeta to create a matrix of all pairwise comparison
This indirect calling is needed since calling functions other than those from psychOpenCama package is blocked on the opencpu server for security reasons.
}
\author{
Robert Studtrucker
}
