% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMetaModel.R
\name{netMetaModel}
\alias{netMetaModel}
\title{netmeta model}
\usage{
netMetaModel(d, reference = "Placebo")
}
\arguments{
\item{d}{A \code{string} representing the dataset name that is used to fit the netmeta model.}

\item{reference}{A \code{string} representing the reference category that should be used for the netmeta model.}
}
\value{
returns nothing but creates a netmeta model which could be uses in other function calls
}
\description{
Creates a netmeta model which is used for further function calls
}
\author{
Robert Studtrucker
}
