% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMetaSummary.R
\name{netMetaSummary}
\alias{netMetaSummary}
\title{Summary for netmeta model}
\usage{
netMetaSummary(model)
}
\arguments{
\item{model}{A netmeta model object}
}
\value{
returns a summary for a given netmeta model.
}
\description{
Using the R package netmeta to to create and return a summary for a given netmeta model.
This indirect calling is needed since calling functions other than those from psychOpenCama package is blocked on the opencpu server for security reasons.
}
\author{
Robert Studtrucker
}
