% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcurves.R
\name{pcurves}
\alias{pcurves}
\title{Plot pcurve}
\usage{
pcurves(yi, vi, measure, d)
}
\arguments{
\item{yi}{A \code{string} of the variable which holds the vector of length k with the observed effect sizes or outcomes in the selected dataset (d)}

\item{vi}{A \code{string} of the variable which holds the vector of length k with the corresponding sampling variances in the selected dataset (d)}

\item{measure}{A character string indicating underlying summary measure}

\item{d}{A \code{string} representing the dataset name that should be used for fitting.}
}
\value{
returns a pcurve plot
}
\description{
Using the R package meta to calculate a random effects meta-analysis based on estimates (e.g. log hazard ratios) and their standard errors. Afterwards using dmetar package to create a pcurve plot.
}
\author{
Robert Studtrucker
}
