% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmaModel.R
\name{rmaModel}
\alias{rmaModel}
\title{rma Model}
\usage{
rmaModel(yi, vi, measure, d, pred1 = NULL, pred2 = NULL)
}
\arguments{
\item{yi}{A \code{string} of the variable which holds the vector of length k with the observed effect sizes or outcomes in the selected dataset (d)}

\item{vi}{A \code{string} of the variable which holds the vector of length k with the corresponding sampling variances in the selected dataset (d)}

\item{measure}{A character string indicating underlying summary measure}

\item{d}{A \code{string} representing the dataset name that should be used for fitting.}

\item{pred1}{Optional parameter of type \code{String}, wich represents the name of the variable which holds the vectors used as input for the metafor mods argument in the selected dataset (d).}

\item{pred2}{Optional parameter of type \code{String}, wich represents the name of the variable which holds the vectors used as input for the metafor mods argument in the selected dataset (d).}
}
\value{
returns a fitted rma_uni model
}
\description{
Using metafor rma.uni function to fit and return a meta-analytic univariate fixed- and random/mixed-effects model with or without moderators for the given dataset.
See Documentation of metafor package for details.
}
\author{
Robert Studtrucker
}
