% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Scatter Plot}
\usage{
scatterPlot(yi, d, pred1 = NULL, pred2 = NULL, effectName = "Effect")
}
\arguments{
\item{d}{A \code{string} representing the dataset name.}

\item{pred1}{A \code{list} for first predictor containing entries for type and value.}

\item{pred2}{A \code{list} for first predictor containing entries for type and value.}
}
\value{
returns a scatterplot for the given dataset and predictor variables.
}
\description{
Produces a scatterplot depending of given dataset and optionally predictors.
}
\author{
Robert Studtrucker
}
